+++ src/client/main.cpp
@@ -126,7 +126,7 @@ int main(int argc, char** argv)
             QStringList data_files;
             QJsonObject job;
             JsonHandler::ReadJsonFile(job, str);
-
+            qDebug() << job;
             if (job.keys().contains("main")) {
 
                 Simulator* simulator = new Simulator;